refactor: moving src to lib
