+++ lib/GARSUtils.ts
