+++ lib/grid/BandLettersRange.ts
