+++ lib/grid/BandNumberRange.ts
