+++ lib/grid/GridTypeUtils.ts
