+++ lib/property/GARSProperties.ts
