+++ package.json
@@ -28,7 +28,7 @@
     "main": "index.ts",
     "dependencies": {
         "@ngageoint/color-js": "^2.0.1",
-        "@ngageoint/grid-js": "^0.0.1-beta.12",
+        "@ngageoint/grid-js": "^0.0.1-beta.13",
         "@ngageoint/simple-features-js": "1.1.1-beta.0",
         "winston": "^3.8.2"
     },
@@ -63,6 +63,9 @@
     "files": [
         "dist"
     ],
+    "directories": {
+        "lib": "lib"
+    },
     "nyc": {
         "temp-dir": "./.test_run",
         "report-dir": "./docs/coverage",
